zoning/requirements1.txt
asn1crypto==0.24.0
bcrypt==3.1.4
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
cryptography==2.1.4
DateTime==4.2
Django==1.11.18
django-auth-ldap==1.1
django-datatables-server-side==0.1
django-guardian==1.4.1
django-python3-ldap==0.11.2
django-queryset-csv==1.0.1
django-userena==2.0.1
djangorestframework==3.9.0
docopt==0.6.2
easy-thumbnails==2.5
enum34==1.1.6
Flask==0.12.2
Flask-MySQL==1.4.0
functools==0.5
html2text==2014.12.29
hurry==1.0
idna==2.6
ipaddress==1.0.19
itsdangerous==0.24
Jinja2==2.10
ldap3==2.5
logging==0.4.9.6
MarkupSafe==1.0
mysql-connector==2.1.4
mysql-connector-python==2.1.7
mysql-python==1.2.5
paramiko==2.4.0
Pillow==5.1.0
pipreqs==0.3.5
prompt-toolkit==2.0.8
py==1.4.34
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
pyflakes==2.1.0
Pygments==2.3.1
pyldap==2.4.45
PyMySQL==0.8.0
PyNaCl==1.2.1
pypiwin32==223
python-ldap==2.4.7
python3-ldap==0.9.8.4
pytz==2018.3
PyU4V==1.0.2
pyvim==2.0.24
pywin32==223
requests==2.18.4
simplejson==3.16.0
six==1.11.0
unicodecsv==0.14.1
UNKNOWN==0.0.0
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
wfastcgi==3.0.0
yarg==0.1.9
zope.interface==4.4.3
